makes possible to override byte format
